update log for form
